Open room for types of reports
